In the class CxxFloatLiteral the method getValueImpl does not work without ".doubleValue()" in the end

+++ ClavaWeaver/src/pt/up/fe/specs/clava/weaver/joinpoints/CxxFloatLiteral.java
@@ -34,8 +34,7 @@ public class CxxFloatLiteral extends AFloatLiteral {
 
     @Override
     public Double getValueImpl() {
-        return literal.get(FloatingLiteral.VALUE);
+        return literal.get(FloatingLiteral.VALUE).doubleValue();
     }
-
 
 }
